+++ src/lib/bm_bsortX/c/common/javax_rtcbench_RTCBenchmark.c
@@ -1,11 +1,11 @@
 #include <stdint.h>
 #include "config.h"
 
-void __attribute__((noinline)) rtcbenchmark_measure_native_performance(uint16_t NUMNUMBERS, int16_t numbers[]);
+void __attribute__((noinline)) rtcbenchmark_measure_native_performance(uint16_t NUMNUMBERS, uint32_t numbers[]);
 
 void javax_rtcbench_RTCBenchmark_void_test_native() {
 	uint16_t NUMNUMBERS = 256;
-	int16_t numbers[NUMNUMBERS];
+	uint32_t numbers[NUMNUMBERS];
 
 	// Fill the array
 	for (uint16_t i=0; i<NUMNUMBERS; i++)
