+++ src/lib/bm_bsortX/java/javax/rtcbench/RTCBenchmark.java
@@ -8,7 +8,7 @@ public class RTCBenchmark {
     public static String name = "BUBBLESORT NOT OPTIMISED";
     public static native void test_native();
     public static boolean test_java() {
-		short numbers[] = new short[NUMNUMBERS]; // Not including this in the timing since we can't do it in C
+		int numbers[] = new int[NUMNUMBERS]; // Not including this in the timing since we can't do it in C
 
 		// Fill the array
 		for (int i=0; i<NUMNUMBERS; i++)
@@ -26,13 +26,13 @@ public class RTCBenchmark {
         return true;
 	}
 
-	public static void rtcbenchmark_measure_java_performance(short[] numbers) {
+	public static void rtcbenchmark_measure_java_performance(int[] numbers) {
 		RTC.startBenchmarkMeasurement_AOT();
 
 		for (short i=0; i<NUMNUMBERS; i++) {
 			for (short j=0; j<NUMNUMBERS-i-1; j++) {
 				if (numbers[j]>numbers[j+1]) {
-					short temp = numbers[j];
+					int temp = numbers[j];
 					numbers[j] = numbers[j+1];
 					numbers[j+1] = temp;
 				}
